Add the ArchiveAPI, implementation, and integration tests
I had to add the converter factory because the default deserializer only
does response bodies. I need a raw response because I also want to
examine the headers to determine when to stop paging. Buuuut the
provided scalar converters do not do boxed arrays. Thankfully the
bytearray value is _super_ simple to implement.